requirements.txt
amqp==5.0.6
astroid==2.4.2
asyncio==3.4.3
asyncpg==0.23.0
autopep8==1.5.5
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bokeh==2.4.1
bs4==0.0.1
cachetools==4.2.2
celery==5.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
client-python==0.2.0
colorama==0.4.4
crypto==1.4.1
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.6
dateparser==1.0.0
decorator==5.1.0
deepdiff==5.6.0
dfdatetime==20210509
dfvfs==20210918
dnsdb-python==1.2.4
dnspython==2.0.0
dtfabric==20210731
Flask==1.1.2
Flask-Classful==0.14.2
Flask-WTF==0.14.3
future==0.18.2
google-auth==1.34.0
idna==2.10
ipwhois==1.2.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsondiff==1.3.0
jsonify==0.5
kiwisolver==1.3.1
kombu==5.2.1
kube-hunter==0.6.3
kubernetes==12.0.1
lazy-object-proxy==1.4.3
libmagic==1.0
libsmdev-python==20210418.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==8.6.0
Naked==0.1.31
netaddr==0.8.0
netifaces==0.11.0
numpy==1.19.5
oauthlib==3.1.1
ordered-set==4.0.2
packaging==21.0
pandas==1.2.1
paramiko==2.7.2
pathvalidate==2.5.0
pefile==2021.9.2
Pillow==8.1.1
pluggy==1.0.0
prettytable==2.2.1
prompt-toolkit==3.0.22
proxmoxer==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pylint==2.6.2
pymongo==3.12.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
python-magic==0.4.22
python-magic-bin==0.4.14
pytsk3==20210801
pytz==2020.5
pywin32==301
PyYAML==5.4.1
redis==3.5.3
regex==2021.9.24
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scapy==2.4.5
shellescape==3.8.1
six==1.15.0
soupsieve==2.2
thehive4py==1.8.1
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.3
validators==0.18.2
vine==5.0.0
volatility3==1.0.1
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==1.0.1
WMI==1.5.1
wrapt==1.12.1
WTForms==2.3.3
XlsxWriter==1.3.7
xlwt==1.3.0
yara==1.7.7
